### 2025-05-20 #8220 @rkurchin
@danielbalzer, hi there! I'll be editing your JOSS submission, and before I start seeking out reviewers, I have a suggestion for you: Have you thought about making your software into a [Julia package](https://pkgdocs.julialang.org/v1/creating-packages/) and registering it in the General registry? While not strictly required for acceptance to JOSS, it does make a number of things easier, including dependency checking/management and installation by potential users. I notice you also don't currently have any tests built into the package; some testing (preferably automated tests, but at least steps to manually verify that the code is working as intended) is required for JOSS, and structuring your code in the standard way also enables easy running of built-in tests by the package manager.
