### 2025-05-20 #8220 @rkurchin
👋 @hisacro, @jiweiqi, and/or @Arrondissement5etDemi, would you be willing to review this submission for JOSS? We carry out our checklist-driven reviews here in GitHub issues and follow these guidelines: https://joss.readthedocs.io/en/latest/review_criteria.html
