+++ shared_functions/approximating_radii_functions.jl
@@ -105,15 +105,13 @@ function charge_approximating_radii(dimension::Integer,N::Integer,disorder::Numb
     #Assign energies to every site from a Gaussian distribution with standard deviation given by the disorder.
     site_energies = disorder.*randn(N^dimension)
 
-    #Creating lists to track rate sums and approximating radii at each increase of hopping_radius.
+    #Creating lists to track rate sums and hopping_radius at each increase of hopping_radius.
     total_hopping_rate_sum = [0.0]
-    hamiltonian_radii = [0]
     hopping_radii = [0]
 
     #Start hopping and hamiltonian radius at 1.
     hopping_radius = 1
     hamiltonian_radius = 1
-    current_hamiltonian_radius = hamiltonian_radius
 
     #Defining the initial position of the charge, in the middle of the lattice.	
     current_location = ones(dimension) .* (N/2)
@@ -134,15 +132,26 @@ function charge_approximating_radii(dimension::Integer,N::Integer,disorder::Numb
     #Calculating distance from every state to the current state.
     distances_squared = sum([(centres[i] .-  current_location[i]).^2 for i in 1:dimension])
 
+    #Create an empty vector to store hamiltonian information at each increase in hamiltonian radius, and then store the current hamiltonian information.
+    stored_hamiltonian_information = Vector{Any}(zeros(Int(round(N/2))))
+    stored_hamiltonian_information[hamiltonian_radius] = [H,Ht,r,transformed_coupling,evals,evecs,centres,current_state,current_location,distances_squared]
+    
     #Increase hopping_radius until the rate sum converges to a chosen accuracy.
     for a = 1:Int(round(N/2))
 
+        #Increase hopping radius.
         hopping_radius += 1
+        
+        #Save the total hopping rate sum from previous increment.
         hopping_rate_sum = [total_hopping_rate_sum[end]]
+
+        #Creating vector to track hamiltonian radii at each increase of hopping_radius.
+        hamiltonian_radii = [hamiltonian_radius]
 
         #Increase hamiltonian_radius until the rate sum to states within current hopping_radius converges to a chosen accuracy.
         for b = 1:Int(round(N/2))
 
+            #Increase exciton hamiltonian radius.
             if b != 1
                 hamiltonian_radius += 1
             elseif b == 1 && hopping_radius > hamiltonian_radius
@@ -151,21 +160,23 @@ function charge_approximating_radii(dimension::Integer,N::Integer,disorder::Numb
 
             #Checking whether the Hamiltonian got bigger than the predetermined system size N.
             if hamiltonian_radius >= N/2 - 1 
-                if total_hopping_rate_sum[end] != 0
+                if hopping_rate_sum[end] != 0
                     @warn "Convergence not reached while calculating approximating radii. Increase the length of the system (N) to avoid this error."
                 end 
                 return [0, 0]
             end
 
-            #If the hamiltonian is too small, set up and rediagonalise a larger subset of the Hamiltonian.
-            if hamiltonian_radius > current_hamiltonian_radius
-                current_hamiltonian_radius = hamiltonian_radius
+            #If the hamiltonian information at current hamiltonian_radius hasn't already been calculated and stored, calculate and store it.
+            if stored_hamiltonian_information[hamiltonian_radius] == 0
                 H,Ht,r,transformed_coupling = setup_hamiltonian.current_charge_transport_hamiltonian(dimension,N,site_energies,electronic_coupling,bath_reorganisation_energy,kappa,hamiltonian_radius,current_location)
                 evals,evecs = eigen(Ht)
[…]
