### 2025-05-28 #8220 @rkurchin
👋 @mizu-bai, and/or @myousefi2016, would you be willing to review this submission for JOSS? We carry out our checklist-driven reviews here in GitHub issues and follow these guidelines: https://joss.readthedocs.io/en/latest/review_criteria.html

(also, @danielbalzer, if you have other suggestions for reviewers, please let me know! Don't tag them here, though, I'll do that if we end up needing to invite more people, that avoids spamming GitHub notifications)
