### 2025-06-06 #8220 @rkurchin
I certainly appreciate the challenges of high-cost CI. However, separate from automating testing, I would still suggest considering making the code into a package to facilitate installation. And if you really don't want to do that, at the very least, I would include a [`Project.toml`](https://pkgdocs.julialang.org/v1/toml-files/) so that the environment/dependencies are specified in a more standard way than what you're currently doing [here](https://github.com/kassalgroup/dKMC.jl/blob/main/shared_functions/package_loading.jl). This allows you to specify versions of dependencies and avoid your code breaking in the future, and allow users to more easily install everything necessary using the standard workflow. In addition, it would more clearly satisfy this review criterion:

> Installation instructions: Is there a clearly-stated list of dependencies? Ideally these should be handled with an automated package management solution.

In your case, you have specified package dependencies through your ad-hoc `package_loading.jl` file, but not any of their versions, which means that future breaking changes to those packages could in principle break your code, too.

I also notice that you have this [`constants.jl`](https://github.com/kassalgroup/dKMC.jl/blob/main/shared_functions/constants.jl) file and wanted to make you aware of [this package](https://github.com/JuliaPhysics/PhysicalConstants.jl) that you could just depend on to slim things down instead of reimplementing all of those yourself (at lower precision). There is also [Unitful.jl](https://painterqubits.github.io/Unitful.jl/stable/) that can be useful for this sort of thing, but depending on your needs can occasionally be more trouble than it's worth.
