### 2025-06-11 58e2cb8
Include Project.toml and .gitignore
