+++ charge_transport/dKMC_charge_transport_functions.jl
@@ -207,7 +207,7 @@ function dKMC_charge_transport(dimension::Integer,N::Integer,site_energies::Vect
     current_location = ones(dimension).* (N/2)
 
     #Calculate the original system Hamiltonian, the polaron-trandformed system Hamiltonian, postion vectors, and coupling vector.
-    H,Ht,r,transformed_coupling = setup_hamiltonian.current_charge_transport_hamiltonian(dimension,N,site_energies,electronic_coupling,bath_reorganisation_energy,kappa,hamiltonian_radius,current_location)
+    H,Ht,r,transformed_coupling,site_indexes = setup_hamiltonian.current_charge_transport_hamiltonian(dimension,N,site_energies,electronic_coupling,bath_reorganisation_energy,kappa,hamiltonian_radius,current_location)
 
     #Calculating eigenvectors and eigenvalues in the energy eigenbasis.
     evals,evecs = eigen(Ht)
@@ -239,10 +239,12 @@ function dKMC_charge_transport(dimension::Integer,N::Integer,site_energies::Vect
         end
 
         #Rediagonalise a new subset of the Hamiltonian.
-        H,Ht,r,transformed_coupling = setup_hamiltonian.current_charge_transport_hamiltonian(dimension,N,site_energies,electronic_coupling,bath_reorganisation_energy,kappa,hamiltonian_radius,current_location)
+        previous_eigenstate = evecs[:,current_state]
+        previous_site_indexes = copy(site_indexes)
+        H,Ht,r,transformed_coupling,site_indexes = setup_hamiltonian.current_charge_transport_hamiltonian(dimension,N,site_energies,electronic_coupling,bath_reorganisation_energy,kappa,hamiltonian_radius,current_location)
         evals,evecs = eigen(Ht)
         centres = [[evecs[:,i]' * Diagonal(r[:,j]) * evecs[:,i] for i=eachindex(evals)] for j in 1:dimension]
-        current_state = argmin(sum([(centres[i] .- current_location[i]).^2 for i in 1:dimension]))
+        current_state = argmax(((previous_eigenstate[findall(x->x in site_indexes,previous_site_indexes)]' * evecs[findall(x->x in previous_site_indexes,site_indexes),:])[:]).^2)
         current_location = [centres[i][current_state] for i in 1:dimension]
 
         #Finding which states are accessible from the current state.
