+++ shared_functions/setup_hamiltonian.jl
@@ -27,6 +27,7 @@ Creates a hamiltonian for a subset of sites within a hamiltonian radius of the c
 - `Ht`: The polaron-transformed system hamiltonian for the subset of sites. 
 - `r`: Matrix containing the position of every site in the hamiltonian.
 - `transformed_coupling`: Matrix containing only polaron-transformed coupling terms, used for dKMC rate calculations.
+- `site_indexes`: Vector containing the indexes of each site included in the Hamiltonian subset.
     
 """
 function current_charge_transport_hamiltonian(dimension::Integer,N::Integer,site_energies::Vector{<:AbstractFloat},electronic_coupling::Number,bath_reorganisation_energy::Number,kappa::AbstractFloat,hamiltonian_radius::Number,current_location::Vector{<:AbstractFloat})
@@ -53,7 +54,10 @@ function current_charge_transport_hamiltonian(dimension::Integer,N::Integer,site
             end
         end
     end
-    
+
+    #Create a list of indexes to all sites within the hamiltonian subset.
+    site_indexes = location_to_index.(dimension,N,site_locations)
+
     #Predefine the size of hamiltonian matrix and position vectors.
     hamiltonian_size = length(site_locations)
     H = zeros(hamiltonian_size,hamiltonian_size)
@@ -63,7 +67,7 @@ function current_charge_transport_hamiltonian(dimension::Integer,N::Integer,site
     for (i,site_location) in enumerate(site_locations)
 
         #Fill the diagonal elements of H with the appropriate site energy.
-        H[i,i] = site_energies[location_to_index(dimension,N,site_location)]
+        H[i,i] = site_energies[site_indexes[i]]
         
         #Fill the position matrix with coordinates of the site.
         r[i,:] = site_location
@@ -90,7 +94,7 @@ function current_charge_transport_hamiltonian(dimension::Integer,N::Integer,site
         Ht[i,i] = H[i,i] - bath_reorganisation_energy
     end
 
-    return H,Ht,r,transformed_coupling
+    return H,Ht,r,transformed_coupling,site_indexes
 
 end
 
@@ -253,6 +257,7 @@ Creates a hamiltonian for a subset of exciton sites within an exciton hamiltonia
 - `Ht`: The polaron-transformed system hamiltonian for the subset of sites. 
 - `r`: Matrix containing the position of every site in the hamiltonian.
 - `transformed_coupling`: Matrix containing only polaron-transformed coupling terms, used for dKMC rate calculations.
+- `site_indexes`: Vector containing the indexes of each site included in the Hamiltonian subset.
 
 """
 function current_exciton_transport_hamiltonian(dimension::Integer,N::Integer,exciton_site_energies::Vector{<:AbstractFloat},dipole_orientations::Matrix{<:AbstractFloat},transition_dipole_moment::Number,epsilon_r::Number,bath_reorganisation_energy::Number,kappa::AbstractFloat,site_spacing::Number,exciton_hamiltonian_radius::Number,current_location::Vector{<:AbstractFloat})
@@ -262,27 +267,26 @@ function current_exciton_transport_hamiltonian(dimension::Integer,N::Integer,exc
 
     #Create a list of the coordinates of sites that are within the hamiltonian radius.
     site_locations = Vector{Integer}[]
-    site_indexes = Integer[]
     if dimension == 1
         for x in accessible_coordinates[1]
             push!(site_locations,[x])
-            push!(site_indexes,location_to_index(dimension,N,[x]))
         end
     elseif dimension == 2
         for x in accessible_coordinates[1], y in accessible_coordinates[2] 
[…]

[…]
