### 2025-06-11 1033e06
Restructure to enable package installation.
