+++ Project.toml
@@ -1,7 +1,7 @@
 name = "dKMC"
 uuid = "2d8c49de-ef21-4537-b730-3af6bb8153b6"
 authors = ["Daniel Balzer"]
-version = "1.0"
+version = "1.0.0"
 
 [deps]
 Dates = "ade2ca70-3891-5945-98fb-dc099432e06a"
@@ -12,6 +12,7 @@ LinearAlgebra = "37e2e46d-f89d-539d-b4ee-838fcccc9c8e"
 Logging = "56ddb016-857b-54e1-b83d-db4d58db5568"
 Measurements = "eff96d63-e80a-5855-80a2-b1b0885c5ab7"
 NumericalIntegration = "e7bfaba1-d571-5449-8927-abc22e82249b"
+Pkg = "44cfe95a-1eb2-52ea-b672-e2afdf69b78f"
 Polynomials = "f27b6e38-b328-58d1-80ce-0feddd5e7a45"
 QuadGK = "1fd47b50-473d-5c70-9696-f719f8f3bcdc"
 Random = "9a3f8284-a2c9-5f02-9a11-845980a1fd5c"
@@ -28,6 +29,7 @@ LinearAlgebra = "1.11.0"
 Logging = "1.11.0"
 Measurements = "2.12.0"
 NumericalIntegration = "0.3.4"
+Pkg = "1.11.0"
 Polynomials = "4.0.19"
 QuadGK = "2.11.2"
 Random = "1.11.0"
