### 2025-06-17 #8220 @danielbalzer
@rkurchin the code is now formatted as a package to facilitate easier installation of the dependencies.

Although the PhysicalConstants.jl package looks great, the performance cost of keeping units attached to the constants means that it won't make sense for us to use here.
