### 2025-07-02 b4c77ff
Update setup_hamiltonian.jl
