### 2025-07-22 38d9687
Include compute_centres() function
