### 2025-08-04 #8722 @editorialbot
[REVIEW]: dKMC: Delocalised kinetic Monte Carlo for simulating fundamental transport processes involving partially delocalised carriers in disordered materials
**Submitting author:** @danielbalzer (<a href="http://orcid.org/0000-0002-4640-1059">Daniel Balzer</a>)
**Repository:** https://github.com/kassalgroup/dKMC.jl
**Branch with paper.md** (empty if default branch): JOSS_branch
**Version:** v1.0
