### 2025-10-04 #8722 @Arrondissement5etDemi
Hi Prof. Kurchin, I'm so sorry! My email got discontinued and I didn't receive the reminders until they are forwarded to my alumnus email... Will do the review ASAP this weekend!
