### 2025-10-04 #8722 @Arrondissement5etDemi
Hi @danielbalzer, @rkurchin
The paper looks good!
Several comments:
1. please include references in your paper using the citation syntax.
2. This does not look like a Julia package that has functions that is ready to go right after installation. Given the current documentation, I had to navigate to the directories in `src/` to run the code and load dependencies manually... Maybe I'm missing something. Better packaging or clearer instructions in the main readme.md would be appreciated.
3. There is no `test/` directory so technically I cannot check that box. I understand that unit testing may not be ideal to test such a large-scale simulation software. However, please add a declaration of conditions under which you assume that the package is working fine.
4. Please add an estimate of how long we expect the simulations to run in each module with some given computational resources.
5. Please add "guidelines for third parties wishing to 1. Contribute to the software 2. Report issues or problems with the software 3. Seek support" to the readme.md
6. I got the following error when running
 `julia run_dKMC_charge_generation.jl charge_generation_input.in charge_generation_out.out`
