### 2025-10-17 aa10718
Update to how boundary warning is displayed.
