+++ src/exciton_transport/run_dKMC_exciton_transport.jl
@@ -60,7 +60,7 @@ end
 @everywhere include("dKMC_exciton_transport_functions.jl")
 
 #Run the dKMC exciton transport calculations.
-diffusion_coefficient,sampling_time_displacements,sampling_time_energies,sampling_time_IPRs,total_hops = dKMC_exciton_transport_functions.dKMC_exciton_transport_results(inputs["dimension"],inputs["N"],inputs["exciton_disorder"],inputs["transition_dipole_moment"],inputs["epsilon_r"],inputs["exciton_bath_reorganisation_energy"],inputs["exciton_bath_cutoff_energy"],inputs["T"],inputs["site_spacing"],inputs["landscape_iterations"],inputs["trajectory_iterations"],inputs["accuracy"],inputs["end_time"],inputs["number_of_sampling_times"])
+diffusion_coefficient,sampling_time_displacements,sampling_time_energies,sampling_time_IPRs,total_hops,reached_boundary_proportion = dKMC_exciton_transport_functions.dKMC_exciton_transport_results(inputs["dimension"],inputs["N"],inputs["exciton_disorder"],inputs["transition_dipole_moment"],inputs["epsilon_r"],inputs["exciton_bath_reorganisation_energy"],inputs["exciton_bath_cutoff_energy"],inputs["T"],inputs["site_spacing"],inputs["landscape_iterations"],inputs["trajectory_iterations"],inputs["accuracy"],inputs["end_time"],inputs["number_of_sampling_times"])
 
 #Record results and timer output to the output file.
 print(output,"""
@@ -82,7 +82,18 @@ Mean IPRs at sampling times: $(sampling_time_IPRs[1,:])
 Standard errors of the mean IPRs at sampling times: $(sampling_time_IPRs[2,:])
 
 Mean number of hops: $(total_hops[1]) ± $(total_hops[2])
+""")
 
+if reached_boundary_proportion > 0.01
+    print(output,"""
+    ────────────────────────────────────────────────────────────────────────────────
+
+    WARNING: "A large proportion ($(reached_boundary_proportion)) of trajectories terminated early as a charge carrier got too close to the edge of the system. Consider increasing N and repeating the calculation."
+
+    """)
+end
+
+print(output,"""
 ────────────────────────────────────────────────────────────────────────────────
 
 Evaluation completed successfully.
