### 2025-10-22 9e729cf
Include package activation, runtimes, and welcome/finished messages.
