+++ src/charge_generation/run_dKMC_charge_generation.jl
@@ -1,3 +1,10 @@
+#Print a message to indicate the calculation is running.
+println("Welcome to dKMC, the calculation is now running.")
+
+#Activate the dKMC package.
+using Pkg
+Pkg.activate(joinpath(@__DIR__, "..", ".."), io=devnull)
+
 #Loading the required packages.
 include("../shared_functions/package_loading.jl")
 
@@ -28,6 +35,9 @@ CT_lifetimes = [inputs["donor_CT_lifetime"], inputs["interfacial_CT_lifetime"], 
 bath_reorganisation_energies = [inputs["electron_bath_reorganisation_energy"], inputs["hole_bath_reorganisation_energy"], inputs["exciton_bath_reorganisation_energy"]]
 bath_cutoff_energies = [inputs["electron_bath_cutoff_energy"], inputs["hole_bath_cutoff_energy"], inputs["exciton_bath_cutoff_energy"]] 
 
+#Record the starting time.
+start_time = DateTime(now())
+
 #Record function introduction and input file to output file.
 output = open(output_file, "w");
 print(output,"""
@@ -56,7 +66,7 @@ organic photovoltaics, even with little to no energetic offset. Chemical Science
 ────────────────────────────────────────────────────────────────────────────────
 
 Julia version:  $VERSION
-Start time:     $(DateTime(now()))
+Start time:     $start_time
 
 ────────────────────────────────────────────────────────────────────────────────
 User input:
@@ -81,7 +91,7 @@ print(output,"""
 ────────────────────────────────────────────────────────────────────────────────
 Results:
 ────────────────────────────────────────────────────────────────────────────────
-All uncertanties included below are standard errors of the mean.
+All uncertainties included below are standard errors of the mean.
 
 Internal quantum efficiency (IQE): $(mean_outcomes[1,1]+mean_outcomes[1,2]) ± $(sqrt(mean_outcomes[2,1]^2 + mean_outcomes[2,2]^2))
 
@@ -171,12 +181,19 @@ if mean_outcomes[1,7] > 0.01
     """)
 end
 
+#Record the end time and run time and print to output.
+end_time = DateTime(now())
+run_time = canonicalize(end_time - start_time)
 print(output,"""
 ────────────────────────────────────────────────────────────────────────────────
 
 Evaluation completed successfully.
-End time:       $(DateTime(now()))
+End time:   $end_time
+Run time:   $run_time 
 
 ────────────────────────────────────────────────────────────────────────────────
 """)
 close(output)
+
+#Print a message to indicate the calculation is finished.
+println("The dKMC calculation is finished.")
