### 2025-11-05 #8722 @Arrondissement5etDemi
Thanks! I think everything works now except that the references are not under the references section in paper.md. If you could correct this that would be great!
