### 2025-11-13 17a2fb6
Updated README, E_step, and E_limit.
