+++ src/shared_functions/dKMC_hopping_rates.jl
@@ -77,11 +77,11 @@ but only uses a subset of sites that are relevant to each state.
 function charge_transport_dKMC_rate(initial_state::Integer,final_state::Integer,transformed_coupling::Matrix{<:AbstractFloat},evals::Vector{<:AbstractFloat},evecs::Matrix{<:AbstractFloat},K_tot::Matrix{ComplexF64},E_step::AbstractFloat,E_limit::Integer,initial_state_relevant_sites::Vector{<:Integer},final_state_relevant_sites::Vector{<:Integer})
 
     #Start the damping rate at zero.
-    gamma = 0
+    gamma = 0.0
 
     #Check whether energy lies outside the range of precalculated K values.
     E = evals[initial_state] - evals[final_state]
-    if abs(E) < E_limit
+    if abs(E) < E_limit - E_step
 
         #Find the K value (at each of the 5 lambda values) at an energy that is closest to the precomputed values stored in K_tot.
         K = real.(K_tot[Int(round((E+E_limit)/E_step))+1,:])
@@ -102,7 +102,7 @@ function charge_transport_dKMC_rate(initial_state::Integer,final_state::Integer,
         end
     end
 
-    rate = (2/(constants.hbar^2))*real(gamma)
+    rate = (2/(constants.hbar^2))*gamma
 
     return rate
 
@@ -137,11 +137,11 @@ but only uses a subset of sites that are relevant to each state.
 function exciton_transport_dKMC_rate(initial_state::Integer,final_state::Integer,transformed_coupling::Matrix{<:AbstractFloat},evals::Vector{<:AbstractFloat},evecs::Matrix{<:AbstractFloat},K_tot::Matrix{ComplexF64},E_step::AbstractFloat,E_limit::Integer,initial_state_relevant_sites::Vector{<:Integer},final_state_relevant_sites::Vector{<:Integer})
 
     #Start the damping rate at zero.
-    gamma = 0
+    gamma = 0.0
 
     #Check whether energy lies outside the range of precalculated K values.
     E = evals[initial_state] - evals[final_state]
-    if abs(E) < E_limit
+    if abs(E) < E_limit - E_step
         
         #Find the K value (at each of the 5 lambda values) at an energy that is closest to the precomputed values stored in K_tot.
         K=real.(K_tot[Int(round((E+E_limit)/E_step))+1,:])
@@ -159,7 +159,7 @@ function exciton_transport_dKMC_rate(initial_state::Integer,final_state::Integer
 
     end 
 
-    rate = (2/(constants.hbar^2))*real(gamma)
+    rate = (2/(constants.hbar^2))*gamma
 
     return rate
 
@@ -169,7 +169,7 @@ end
 """
     charge_separation_dKMC_rate(initial_state::Integer,final_state::Integer,
     transformed_coupling::Matrix{<:AbstractFloat},evals::Vector{<:AbstractFloat},evecs::Matrix{<:AbstractFloat},
-    K_tots::Vector{Matrix{ComplexF64}},E_steps::Vector{<:AbstractFloat},E_limits::Vector{<:Integer},
+    K_tots::Vector{Matrix{ComplexF64}},E_step::Number,E_limit::Number,
     initial_state_relevant_sites::Vector{<:Integer},final_state_relevant_sites::Vector{<:Integer},
     electron_index::Vector{<:Integer},hole_index::Vector{<:Integer},bath_index::Matrix{Integer})
 
@@ -183,8 +183,8 @@ but only uses a subset of sites that are relevant to each state.
 - `evals`: Energy eigenvalues of the polaron trasformed system hamiltonian.
 - `evecs`: Energy eigenvectors of the polaron trasformed system hamiltonian.
[…]
