### 2025-11-25 1761c1c
Update paper.bib
