+++ JOSS/paper.bib
@@ -1,4 +1,4 @@
-@Article{Balzer2021,
+@article{Balzer2021,
 author ="Balzer, Daniel and Smolders, Thijs J. A. M. and Blyth, David and Hood, Samantha N. and Kassal, Ivan",
 title  ="{D}elocalised kinetic {M}onte {C}arlo for simulating delocalisation-enhanced charge and exciton transport in disordered materials",
 journal  ="Chemical Science",
