### 2025-11-25 8edcf07
Update paper.bib
