+++ JOSS/paper.bib
@@ -26,7 +26,7 @@ URL = {https://www.science.org/doi/abs/10.1126/sciadv.abl9692}
 @article{Balzer2023,
 author = {Balzer, Daniel and Kassal, Ivan},
 title = {{M}echanism of delocalization-enhanced exciton transport in disordered organic semiconductors},
-journal = {J. Phys. Chem. Lett.},
+journal = {Journal of Physical Chemistry Letters},
 volume = {14},
 number = {8},
 pages = {2155-2162},
