### 2025-11-25 #8722 @danielbalzer
@rkurchin @Arrondissement5etDemi Apologies for that. I believe the references have now been fixed and are rendering under the header.
