### 2025-12-06 00171f9
Update paper.md
