+++ JOSS/paper.md
@@ -24,16 +24,16 @@ bibliography: paper.bib
 ---
 
 # Summary
-The movement of charge and energy is a fundamental process in materials science, underpinning technologies such as solar cells, light-emitting diodes, batteries, and electronics. Transport is well understood in both highly ordered materials (band conduction) and highly disordered ones (hopping conduction). However, in moderately disordered materials—including many organic semiconductors—transport lies in the intermediate transport regime between these well-understood extremes. Accurately modelling intermediate-regime conduction is difficult because describing wavefunction delocalisation requires a fully quantum-mechanical treatment, which is challenging in disordered materials that lack periodicity. We describe delocalised kinetic Monte Carlo (`dKMC`), the first theoretical approach to treat, in three dimensions, all the processes crucial in organic semiconductors: disorder, delocalisation, and polaron formation. As a result, it can treat the intermediate transport regime between band and hopping conduction. `dKMC` reveals that the fundamental physics of transport in moderately disordered materials is that of charges and excitons hopping between partially delocalised electronic states. In this work, we release the `dKMC.jl` package, which contains modules for simulating the fundamental processes of charge and exciton transport as well as charge separation and generation.
+The movement of charge and energy is a fundamental process in materials science, underpinning technologies such as solar cells, light-emitting diodes, batteries, and electronics. Transport is well understood in both highly ordered materials (band conduction) and highly disordered ones (hopping conduction). However, in moderately disordered materials—including many organic semiconductors—transport lies in the intermediate transport regime between these well-understood extremes. Accurately modelling intermediate-regime conduction is difficult because describing wavefunction delocalisation requires a fully quantum-mechanical treatment, which is challenging in disordered materials that lack periodicity. We have previously developed delocalised kinetic Monte Carlo (`dKMC`), the first theoretical approach to treat, in three dimensions, all the processes crucial in organic semiconductors: disorder, delocalisation, and polaron formation. As a result, it can treat the intermediate transport regime between band and hopping conduction. `dKMC` reveals that the fundamental physics of transport in moderately disordered materials is that of charges and excitons hopping between partially delocalised electronic states. In this work, we release the `dKMC.jl` package, which contains modules for simulating the fundamental processes of charge and exciton transport as well as charge separation and generation.
 
 ![Figure 1: While transport mechanisms are well understood in the extremes, coherent band conduction through extended states and incoherent hopping through localised states, they remain poorly understood in the intermediate regime where many organic semiconductors lie. Figure adapted with permission from [@Balzer2021].](intermediate_regime.png)
 
 # Statement of need
 The need for dKMC is demonstrated by its ability to solve two important problems. 
 
-First, `dKMC` is the first computational technique that is able to include all of the processes crucial in organic semiconductors while remaining computationally tractable enough to treat realistic, three-dimensional systems on mesoscopic time and length scales [@Balzer2021]. Before `dKMC`, the difficulty in modelling transport in the intermediate regime had prevented the development of such a theory. Instead, prior approaches either needed to exclude one of the key ingredients (disorder, delocalisation, and polaron formation), which could lead to inaccurate results, or included them all but restricted the application to smaller systems (dimension, time, and length scales). `dKMC` solves these problems, striking a balance between the level of approximation and the size of the system it can simulate. Therefore,`dKMC` provides a simulation tool that can both accurately and efficiently simulate fundamental transport processes involving partially delocalised carriers.
[…]
