### 2025-12-10 0abc73f
Create CITATION.cff
