### 2025-12-10 1baed39
Include JOSS badge and citation
